WIP extracting classes from fix_unattached_media, implementing thread pool, codifying requirements, adding CircleCI
