+++ requirements.txt
@@ -1,3 +1,4 @@
 unittest2==1.1.0
 raven==5.32.0
 mock==2.0.0
+elasticsearch==1.4.0
